rely on fs connection with s3cmd
